feat: initial commit
Context: initial commit before placing on GitHub before publishing to
PyPI and testing real-life use. Basic API as I envisioned for the first
iteration is implemented in principle, but is certainly rough around the
edges and will need testing and bug fixes. This is just to get a copy
online to begin the distribution process.